requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==3.2.2
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.0
Flask==3.1.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
Werkzeug==3.1.4
